tests/requirements.txt
# Testing Requirements for TextToHandwriting Project
# Install with: pip install -r tests/requirements.txt

# Core testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0

# HTTP testing
requests>=2.28.0
httpx>=0.24.0

# Image processing for test fixtures
Pillow>=9.0.0

# Backend testing (if needed)
requests-mock>=1.10.0

# Performance testing
pytest-benchmark>=4.0.0

# Test reporting
pytest-html>=3.1.0
allure-pytest>=2.12.0

# Development helpers
pytest-watch>=4.2.0
pytest-sugar>=0.9.6 